bug fixes, code improvements
Engine Changes:
-adjusted logging
-improved saving/loading
-changes to settings file now reflect changes in game
-checkboxes may now be added to menus
-created a way to easily switch between menus
-buttons may now store function pointers that have an integer parameter
Demo Changes:
-fixed checkboxes not working after game load
-updated readme

+++ src/movable_object.cpp
@@ -329,7 +329,7 @@ movable_object::movable_object(): complex_object()
     speed=1.01f;
     rest_position.set(position);
     rest_rotation=rotation;
-    rally = new point2i((int)rest_position.x,(int)rest_position.y);
+    rally = new point2i(rest_position.x,rest_position.y);
     rally_set=false;
     reset_motion();
 }
